# apis/app-api/community/openapi.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "SDKWork Community App API",
    "version": "1.0.0",
    "x-sdkwork-owner": "sdkwork-community",
    "x-sdkwork-api-authority": "sdkwork-community.app"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:18080"
    }
  ],
...
  "paths": {
    "/app/v3/api/community/categories": {
      "get": {
        "tags": [
          "community"
        ],
        "summary": "Community categories.list",
        "operationId": "categories.list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SdkWorkListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-community",
        "x-sdkwork-api-authority": "sdkwork-community.app",
        "x-sdkwork-domain": "community",
        "x-sdkwork-resource": "categories",
        "x-sdkwork-public": false,
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-standard-profile": "sdkwork-v3"
      }
    },
    "/app/v3/api/community/feed": {
      "get": {
        "tags": [
          "community"
        ],
        "summary": "Community feed.list",
        "operationId": "feed.list",
        "parameters": [
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reviewState",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SdkWorkListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-community",
        "x-sdkwork-api-authority": "sdkwork-community.app",
        "x-sdkwork-domain": "community",
        "x-sdkwork-resource": "feed",
        "x-sdkwork-public": false,
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-standard-profile": "sdkwork-v3"
      }
    },
    "/app/v3/api/community/entries/{entryId}": {
      "get": {
        "tags": [
          "community"
        ],
        "summary": "Community entries.retrieve",
        "operationId": "entries.retrieve",
        "parameters": [
          null
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SdkWorkResourceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-community",
        "x-sdkwork-api-authority": "sdkwork-community.app",
        "x-sdkwork-domain": "community",
        "x-sdkwork-resource": "entries",
        "x-sdkwork-public": false,
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-standard-profile": "sdkwork-v3"
      },
      "patch": {
        "tags": [
          "community"
        ],
        "summary": "Community entries.update",
        "operationId": "entries.update",
        "parameters": [
          null
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommunityEntryCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SdkWorkResourceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-community",
        "x-sdkwork-api-authority": "sdkwork-community.app",
        "x-sdkwork-domain": "community",
        "x-sdkwork-resource": "entries",
        "x-sdkwork-public": false,
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-standard-profile": "sdkwork-v3"
      },
      "delete": {
        "tags": [
          "community"
        ],
        "summary": "Community entries.delete",
        "operationId": "entries.delete",
        "parameters": [
          null
        ],
        "responses": {
          "204": {
            "description": "OK"
          },
          "400": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-community",
        "x-sdkwork-api-authority": "sdkwork-community.app",
        "x-sdkwork-domain": "community",
        "x-sdkwork-resource": "entries",
        "x-sdkwork-public": false,
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-standard-profile": "sdkwork-v3"
      }
    },
    "/app/v3/api/community/entries/{entryId}/recommendations": {
      "get": {
        "tags": [
          "community"
        ],
        "summary": "Community entries.recommendations.list",
        "operationId": "entries.recommendations.list",
        "parameters": [
          null
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SdkWorkListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-community",
        "x-sdkwork-api-authority": "sdkwork-community.app",
        "x-sdkwork-domain": "community",
        "x-sdkwork-resource": "entries",
        "x-sdkwork-public": false,
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-standard-profile": "sdkwork-v3"
      }
    },
    "/app/v3/api/community/entries": {
      "post": {
        "tags": [
          "community"
        ],
        "summary": "Community entries.create",
        "operationId": "entries.create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommunityEntryCommand"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SdkWorkResourceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-community",
        "x-sdkwork-api-authority": "sdkwork-community.app",
        "x-sdkwork-domain": "community",
        "x-sdkwork-resource": "entries",
        "x-sdkwork-public": false,
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-standard-profile": "sdkwork-v3"
      }
    },
    "/app/v3/api/community/entries/{entryId}/publication_readiness": {
      "get": {
        "tags": [
          "community"
        ],
        "summary": "Community entries.publicationReadiness.retrieve",
        "operationId": "entries.publicationReadiness.retrieve",
        "parameters": [
          null
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SdkWorkResourceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-community",
        "x-sdkwork-api-authority": "sdkwork-community.app",
        "x-sdkwork-domain": "community",
        "x-sdkwork-resource": "entries",
        "x-sdkwork-public": false,
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-standard-profile": "sdkwork-v3"
      }
    },
    "/app/v3/api/community/entries/{entryId}/reactions": {
      "post": {
        "tags": [
          "community"
        ],
        "summary": "Community reactions.set",
        "operationId": "reactions.create",
        "parameters": [
          null
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommunityReactionCommand"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SdkWorkResourceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-community",
        "x-sdkwork-api-authority": "sdkwork-community.app",
        "x-sdkwork-domain": "community",
        "x-sdkwork-resource": "reactions",
        "x-sdkwork-public": false,
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-standard-profile": "sdkwork-v3"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AuthToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Access-Token"
      },
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
      "SdkWorkApiResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "data",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "enum": [
              0
            ],
            "default": 0,
            "minimum": 0,
            "maximum": 0
          },
          "data": {
            "description": "Operation-specific payload typed per response schema."
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          }
        }
      },
      "SdkWorkResourceData": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "item"
        ],
        "properties": {
          "item": {
            "type": "object",
            "additionalProperties": true,
            "description": "Typed domain resource for the operation."
          }
        }
      },
      "SdkWorkPageData": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items",
          "pageInfo"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "pageInfo": {
            "$ref": "#/components/schemas/PageInfo"
          }
        }
      },
...
      "SdkWorkPlatformErrorCode": {
        "type": "integer",
        "format": "int32",
        "minimum": 40001,
        "maximum": 79999,
        "description": "Platform or domain error code per API_SPEC.md §15.3."
      },
      "ProblemDetail": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "type",
          "title",
          "status",
          "code",
          "traceId"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "$ref": "#/components/schemas/SdkWorkPlatformErrorCode"
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          },
          "i18nKey": {
            "type": "string",
            "description": "Optional stable localization key such as errors.result.40001."
          },
          "locale": {
            "type": "string",
            "description": "Optional effective BCP 47 locale used by framework message mapping."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        }
      },
...
      "SdkWorkResourceResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/SdkWorkResourceData"
              }
            }
          }
        ]
      },
      "SdkWorkListResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/SdkWorkPageData"
              }
            }
          }
        ]
      },
...
      "CommunityEntryCommand": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "categoryId",
          "kind",
          "title"
        ],
        "properties": {
          "categoryId": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/CommunityEntryKind"
          },
          "title": {
            "type": "string"
          },
          "excerpt": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "CommunityReactionCommand": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "reactionType",
          "active"
        ],
        "properties": {
          "reactionType": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
...
}